requirements.txt
aiogram==3.11.0
Flask==3.0.0
psycopg2-binary==2.9.9
aiohttp==3.9.0